// src/lib/sanity/types.ts
export type GetWorkoutRecordQueryResult = {
  _id: string;
  date: string | null;
  duration: number | null;
  exercises: Array<{
    exercise: {
      _id: string;
      name: string | null;
    } | null;
    sets: Array<{
      reps: number | null;
      weight: number | null;
      weightunit: null;
      _type: "exerciseSet";
      _key: string;
    }> | null;
    _type: "workoutExercise";
    _key: string;
  }> | null;
} | null;